Fix preferences window to center on focused screen
