+++ app/main-es6/server/app/mainwindow.js
@@ -6,6 +6,7 @@ const BrowserWindow = electron.BrowserWindow;
 
 const platformUtil = require('../../../platform-util');
 const rpc = require('../rpc-server');
+const windowUtil = require('./windowutil');
 
 let browserWindow = null;
 
@@ -41,41 +42,12 @@ function createWindow(cb) {
   });
 }
 
-function _centerWindowOnSelectedScreen(window) {
-  const screen = electron.screen;
-
-  let selectedDisplay = screen.getPrimaryDisplay();
-  const displays = screen.getAllDisplays();
-  const cursorPos = screen.getCursorScreenPoint();
-
-  for (const display of displays) {
-    const bounds = display.bounds;
-    const [left, right, top, bottom] = [bounds.x, bounds.x + bounds.width, bounds.y, bounds.y + bounds.height];
-    if (cursorPos.x < left || cursorPos.x > right)
-      continue;
-    if (cursorPos.y < top || cursorPos.y > bottom)
-      continue;
-
-    selectedDisplay = display;
-    break;
-  }
-
-  const windowSize = window.getSize();
-  const displayBounds = selectedDisplay.bounds;
-
-  const centerPos = [displayBounds.x + displayBounds.width * 0.5, displayBounds.y + displayBounds.height * 0.5];
-  centerPos[0] -= windowSize[0] * 0.5; // x
-  centerPos[1] -= windowSize[1] * 0.5; // y
-
-  window.setPosition(Math.round(centerPos[0]), Math.round(centerPos[1]));
-}
-
 function showWindowOnCenter() {
   if (browserWindow === null)
     return;
 
   platformUtil.saveFocus();
-  _centerWindowOnSelectedScreen(browserWindow);
+  windowUtil.centerWindowOnSelectedScreen(browserWindow);
   browserWindow.show();
 }
