+++ app/main-es6/server/app/prefwindow.js
@@ -1,7 +1,9 @@
 'use strict';
 
 const electron = require('electron');
+const shell = electron.shell;
 const BrowserWindow = electron.BrowserWindow;
+const windowUtil = require('./windowutil');
 
 let browserWindow = null;
 
@@ -11,9 +13,7 @@ function show() {
   browserWindow = new BrowserWindow({
     width: 800,
     height: 650,
-    center: true,
-    show: false,
-    maximizable: false
+    show: false
   });
   browserWindow.loadURL(`file://${__dirname}/../../../dist/preferences.html`);
   browserWindow.on('close', () => {
@@ -22,7 +22,14 @@ function show() {
     const server = require('../server');
     server.commitPreferences();
   });
+
+  browserWindow.webContents.on('will-navigate', (evt, url) => {
+    shell.openExternal(encodeURI(url));
+    evt.preventDefault();
+  });
   browserWindow.setMenuBarVisibility(false);
+
+  windowUtil.centerWindowOnSelectedScreen(browserWindow);
   browserWindow.show();
 }
